Changes from review.
