+++ app/models/concerns/client_search.rb
@@ -1,0 +1,77 @@
+###
+# Copyright 2016 - 2022 Green River Data Analysis, LLC
+#
+# License detail: https://github.com/greenriver/hmis-warehouse/blob/production/LICENSE.md
+###
+
+module ClientSearch
+  extend ActiveSupport::Concern
+  included do
+    # Requires a block!
+    def self.text_searcher(text, client_scope:) # rubocop:disable Lint/UnusedMethodArgument
+      return none unless text.present?
+
+      text.strip!
+      sa = arel_table
+      alpha_numeric = /[[[:alnum:]]-]+/.match(text).try(:[], 0) == text
+      numeric = /[\d-]+/.match(text).try(:[], 0) == text
+      date = /\d\d\/\d\d\/\d\d\d\d/.match(text).try(:[], 0) == text
+      social = /\d\d\d-\d\d-\d\d\d\d/.match(text).try(:[], 0) == text
+      # Explicitly search for only last, first if there's a comma in the search
+      if text.include?(',')
+        last, first = text.split(',').map(&:strip)
+        where = sa[:LastName].lower.matches("#{last.downcase}%") if last.present?
+        if last.present? && first.present?
+          where = where.and(sa[:FirstName].lower.matches("#{first.downcase}%"))
+        elsif first.present?
+          where = sa[:FirstName].lower.matches("#{first.downcase}%")
+        end
+        # Explicity search for "first last"
+      elsif text.include?(' ')
+        first, last = text.split(' ').map(&:strip)
+        where = sa[:FirstName].lower.matches("#{first.downcase}%").
+          and(sa[:LastName].lower.matches("#{last.downcase}%"))
+        # Explicitly search for a PersonalID
+      elsif alpha_numeric && (text.size == 32 || text.size == 36)
+        where = sa[:PersonalID].matches(text.gsub('-', ''))
+      elsif social
+        where = sa[:SSN].eq(text.gsub('-', ''))
+      elsif date
+        (month, day, year) = text.split('/')
+        where = sa[:DOB].eq("#{year}-#{month}-#{day}")
+      elsif numeric
+        where = sa[:PersonalID].eq(text).or(sa[:id].eq(text))
+      else
+        query = "%#{text}%"
+        where = sa[:FirstName].matches(query).
+          or(sa[:LastName].matches(query))
+
+        where = nickname_search(where, text)
+        where = metaphone_search(where, :FirstName, text)
+        where = metaphone_search(where, :LastName, text)
+      end
+      begin
+        client_ids = yield(where)
+      rescue RangeError
+        return none
+      end
+
+      client_ids << text if numeric && self.destination.where(id: text).exists? # rubocop:disable Style/RedundantSelf
+      where(id: client_ids)
+    end
+
+    def self.nickname_search(where, text)
+      nicks = Nickname.for(text).map(&:name)
+      where = where.or(nf('LOWER', [arel_table[:FirstName]]).in(nicks)) if nicks.any?
+
+      where
+    end
+
+    def self.metaphone_search(where, field, text)
+      alt_names = UniqueName.where(double_metaphone: Text::Metaphone.double_metaphone(text).to_s).map(&:name)
+      where = where.or(nf('LOWER', [arel_table[field]]).in(alt_names)) if alt_names.present?
+
+      where
+    end
+  end
+end
